[WIP]: 멀티 플레이 UI 구현중 멀티 화면 방 이름, 비밀번호 변수 초기화 로직 추가

+++ YogaYo/app/src/main/java/com/d104/yogaapp/features/multi/MultiScreen.kt
@@ -56,7 +56,7 @@ import com.d104.yogaapp.features.multi.dialog.EnterRoomDialog
 
 @Composable
 fun MultiScreen(
-    onNavigateMultiPlay: (Int) -> Unit,
+    onNavigateMultiPlay: (Long) -> Unit,
     viewModel: MultiViewModel = hiltViewModel()
 ) {
     val uiState by viewModel.uiState.collectAsState()
